test/requirements.txt
pytest==5.4.2
psycopg2-binary==2.8.5
-e ../.
